SCHED-137: Modify Manager structure to encompass run logic
