+++ process_manager/scheduler.py
@@ -8,6 +8,7 @@ from components.collector import Collector
 from components.selector import Selector
 from components.optimizer import Optimizer
 from components.optimizer.dummy import DummyOptimizer
+from common.output import print_plans
 
 
 @dataclass
@@ -35,7 +36,6 @@ class SchedulerConfig:
 class Scheduler():
     def __init__(self, config: SchedulerConfig):
         self.config = config
-        self.plan = None
 
     def __call__(self):
         signal.signal(signal.SIGINT, signal.SIG_IGN)
@@ -54,17 +54,14 @@ class Scheduler():
         collector.load_programs(program_provider=OcsProgramProvider(),
                                 data=programs)
 
-        # selector = Selector(
-        #    collector=collector,
-        #    properties=self.config.selector.properties
-        #)
+        selector = Selector(collector=collector,
+                            properties=self.config.selector.properties)
 
         # Execute the Selector.
         # Not sure the best way to display the output.
-        #selection = selector.select()
+        selection = selector.select()
         # Execute the Optimizer.
-        #dummy = DummyOptimizer()
-        #optimizer = Optimizer(selection, algorithm=dummy)
-        #plans = optimizer.schedule()
-        #self.plans = plans
-        # print_plans(plans)
+        dummy = DummyOptimizer()
+        optimizer = Optimizer(selection, algorithm=dummy)
+        plans = optimizer.schedule()
+        print_plans(plans)
