+++ scripts/run_with_manager.py
@@ -12,52 +12,34 @@ from process_manager.scheduler import Scheduler, SchedulerConfig, CollectorConfi
 from api.observatory.gemini import GeminiProperties
 from common.minimodel import ALL_SITES, Semester, SemesterHalf, ProgramTypes, ObservationClass
 
+logging.basicConfig(level=logging.INFO)
 
-async def main(config: SchedulerConfig, size: int, timeout: int, period: int):
-    done = asyncio.Event()
-    manager = ProcessManager(size, timeout)
+# Create Scheduler
+collector_config = CollectorConfig({Semester(2018, SemesterHalf.B)},
+                                   {ProgramTypes.Q, ProgramTypes.LP, ProgramTypes.FT, ProgramTypes.DD},
+                                   {ObservationClass.SCIENCE, ObservationClass.PROGCAL, ObservationClass.PARTNERCAL}
+                                   )
 
-    def shutdown():
-        done.set()
-        manager.shutdown()
-        asyncio.get_event_loop().stop()
+selector_config = SelectorConfig(GeminiProperties)
 
-    asyncio.get_event_loop().add_signal_handler(signal.SIGINT, shutdown)
+config = SchedulerConfig(Time("2018-10-01 08:00:00", format='iso', scale='utc'),
+                         Time("2018-10-03 08:00:00", format='iso', scale='utc'),
+                         1.0 * u.min,
+                         ALL_SITES,
+                         collector_config,
+                         selector_config)
+                         
+scheduler = Scheduler(config)
 
-    while not done.is_set():
-        
-        # logging.info(f"Scheduling a job for {task}")
-        scheduler = Scheduler(config)
-        manager.add_task(datetime.now(), scheduler, mode, timeout)
-        if period == 0:
-            # random case #
-            await asyncio.sleep(randint(1, 10))
-        else:
-            await asyncio.sleep(period)
+# Manager params
+mode = TaskType.STANDARD  # Type of runner is going to be working
+size = 1  # number of processes
+timeout = 60 * 60  # max time to wait for a process to finish
+period = 2000000
 
-
-if __name__ == '__main__':
-    
-    logging.basicConfig(level=logging.INFO)
-    mode = TaskType.STANDARD
-    collector_config = CollectorConfig({Semester(2018, SemesterHalf.B)},
-                                       {ProgramTypes.Q, ProgramTypes.LP, ProgramTypes.FT, ProgramTypes.DD},
-                                       {ObservationClass.SCIENCE, ObservationClass.PROGCAL, ObservationClass.PARTNERCAL}
-                                       )
-    selector_config = SelectorConfig(GeminiProperties)
-    config = SchedulerConfig(Time("2018-10-01 08:00:00", format='iso', scale='utc'),
-                             Time("2018-10-03 08:00:00", format='iso', scale='utc'),
-                             1.0 * u.min,
-                             ALL_SITES,
-                             collector_config,
-                             selector_config)
-
-    # Manager params
-    size = 1  # number of processes
-    timeout = 60 * 60  # max time to wait for a process to finish
-    period = 2000000
-    try:
-        asyncio.run(main(config, size, timeout, period))
-    except RuntimeError:
-        # Likely you pressed Ctrl-C...
-        ...
+manager = ProcessManager(size, timeout)
+try:
+    asyncio.run(manager.run(scheduler, period, mode))
+except RuntimeError:
+    # Likely you pressed Ctrl-C...
+    ...
